refactor: 회원 차단 기능 application 객체 분리
