+++ src/main/java/dandi/dandi/member/application/port/in/MemberBlockUseCasePort.java
@@ -1,0 +1,6 @@
+package dandi.dandi.member.application.port.in;
+
+public interface MemberBlockUseCasePort {
+
+    void blockMember(Long memberId, MemberBlockCommand memberBlockCommand);
+}
