+++ src/main/java/dandi/dandi/member/application/port/in/MemberUseCaseServicePort.java
@@ -5,6 +5,4 @@ public interface MemberUseCaseServicePort {
     void updateNickname(Long memberId, NicknameUpdateCommand nicknameUpdateCommand);
 
     void updateLocation(Long memberId, LocationUpdateCommand locationUpdateCommand);
-
-    void blockMember(Long memberId, MemberBlockCommand memberBlockCommand);
 }
