+++ src/test/java/dandi/dandi/member/application/service/MemberUseCaseServiceAdapterTest.java
@@ -3,17 +3,13 @@ package dandi.dandi.member.application.service;
 import static dandi.dandi.member.MemberTestFixture.DISTRICT;
 import static dandi.dandi.member.MemberTestFixture.DISTRICT_VALUE;
 import static dandi.dandi.member.MemberTestFixture.MEMBER;
-import static dandi.dandi.member.MemberTestFixture.MEMBER_ID;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import dandi.dandi.common.exception.NotFoundException;
 import dandi.dandi.member.application.port.in.LocationUpdateCommand;
-import dandi.dandi.member.application.port.in.MemberBlockCommand;
 import dandi.dandi.member.application.port.in.NicknameUpdateCommand;
-import dandi.dandi.member.application.port.out.MemberBlockPersistencePort;
 import dandi.dandi.member.application.port.out.MemberPersistencePort;
 import dandi.dandi.member.domain.DistrictParser;
 import dandi.dandi.member.domain.Location;
@@ -29,11 +25,9 @@ import org.springframework.dao.DataIntegrityViolationException;
 class MemberUseCaseServiceAdapterTest {
 
     private final MemberPersistencePort memberPersistencePort = Mockito.mock(MemberPersistencePort.class);
-    private final MemberBlockPersistencePort memberBlockPersistencePort =
-            Mockito.mock(MemberBlockPersistencePort.class);
     private final DistrictParser districtParser = new DistrictParser();
     private final MemberUseCaseServiceAdapter memberService =
-            new MemberUseCaseServiceAdapter(memberPersistencePort, memberBlockPersistencePort, districtParser);
+            new MemberUseCaseServiceAdapter(memberPersistencePort, districtParser);
 
     @DisplayName("회원의 닉네임을 변경할 수 있다.")
     @Test
@@ -75,45 +69,5 @@ class MemberUseCaseServiceAdapterTest {
         memberService.updateLocation(MEMBER.getId(), locationUpdateCommand);
 
         verify(memberPersistencePort).updateLocation(MEMBER.getId(), new Location(latitude, longitude, DISTRICT));
-    }
-
-    @DisplayName("존재하지 않는 id의 사용자를 차단하려고 하면 예외를 발생시킨다.")
-    @Test
-    void blockMember_NotFound() {
-        Long blockedMemberId = 2L;
-        when(memberPersistencePort.existsById(blockedMemberId))
-                .thenReturn(false);
-
-        assertThatThrownBy(() -> memberService.blockMember(MEMBER_ID, new MemberBlockCommand(blockedMemberId)))
-                .isInstanceOf(NotFoundException.class)
-                .hasMessage(NotFoundException.member().getMessage());
-    }
-
-    @DisplayName("이미 차단한 사용자를 차단하려고 하면 예외를 발생시킨다.")
-    @Test
-    void blockMember_AlreadyBlocked() {
-        Long blockedMemberId = 2L;
-        when(memberPersistencePort.existsById(blockedMemberId))
-                .thenReturn(true);
-        when(memberBlockPersistencePort.existsByBlockingMemberIdAndBlockedMemberId(MEMBER_ID, blockedMemberId))
-                .thenReturn(true);
-
-        assertThatThrownBy(() -> memberService.blockMember(MEMBER_ID, new MemberBlockCommand(blockedMemberId)))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessage("이미 차단한 사용자입니다.");
-    }
-
-    @DisplayName("다른 사용자를 차단할 수 있다.")
-    @Test
-    void blockMember() {
-        Long blockedMemberId = 2L;
-        when(memberPersistencePort.existsById(blockedMemberId))
-                .thenReturn(true);
-        when(memberBlockPersistencePort.existsByBlockingMemberIdAndBlockedMemberId(MEMBER_ID, blockedMemberId))
-                .thenReturn(false);
-
-        memberService.blockMember(MEMBER_ID, new MemberBlockCommand(blockedMemberId));
-
-        verify(memberBlockPersistencePort).saveMemberBlockOf(MEMBER_ID, blockedMemberId);
     }
 }
